move jsApi to subdirectory core
