+++ lib/mocks/substrate_api/mockJSApi.dart
@@ -1,3 +1,3 @@
-import 'package:encointer_wallet/service/substrate_api/jsApi.dart';
+import 'package:encointer_wallet/service/substrate_api/core/jsApi.dart';
 
 class MockJSApi extends JSApi {}
